insightproject/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
dynaconf==3.1.4
hvac==0.10.8
idna==2.10
loguru==0.5.3
oauthlib==3.1.0
Pillow==8.1.2
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.3
validators==0.18.2
win32-setctime==1.0.3
